Allow vtwm's own windows to start iconified
Applications go through HandleMapRequest() in events.c which deals
with "initial-state" issues.  In contrast, vtwm's own windows, e.g.
icon managers and the VD-window do not (vtwm maps and raises them
"by itself").

This is annoying indeed, and as now there is someone else seeing
this as a problem, I finally fixed this.  :-)

Here is a patch which now sends icon managers and the VD window
through "usual" startup initial-state management too, so they should
now start iconified if listed in "StartIconfied".  One can even on
icon manager basis specify which ones appear visible on startup, and
which ones hidden.  In fact the "ShowIconManager" would be
superfluous now, but due to backwards compatibility reasons there is
no point to remove it.  As clients usually start in normal state, so
would now icon managers, which would be the opposite of the
historical default (the default seems to be to hide icon managers).
 Without any better idea I implemented the "StartIconified" to take
effect only if "ShowIconManager" is specified too. (If its not, all
icon managers are hidden, which would be then equivalent to put them
all into "StartIconified" anyways.)

I haven't done any heavy testing of this patch yet, though using

    StartIconified { "xv Icon Manager" "VTWM Desktop" }
    ShowIconManager
    DontIconifyByUnmapping  { "VTWM Desktop" }
    IconManagerDontShow  { "VTWM *" }
    IconManagers { "xv" "XVs" "120x20-20-120@1" 1 }

in .vtwmrc seems to work as expected for now.